It works!  Lots of changes here.  Mostly, to add an edit list when a track needs to back up to sync points.  The edit list makes sure that the tracks begin at the same point.

+++ trunk/mp4splitter/src/mp4/util/Mp4Split.java
@@ -34,7 +34,7 @@ public class Mp4Split extends DefaultAtomVisitor {
   public static String inputFile;
   public static String outputFile;
   public static boolean mdat = true;
-  public static long time;
+  public static float time;
   
   @Override
   protected void defaultAction(Atom atom) throws AtomException {
@@ -179,7 +179,7 @@ public class Mp4Split extends DefaultAtomVisitor {
         outputFile = args[++i];
       }
       else if (arg.equals("-time")) {
-        time = Long.valueOf(args[++i]);
+        time = Float.valueOf(args[++i]);
       }
       else if (arg.equals("-no_mdat")) {
         mdat = false;
